+++ SampleApps/PingExample/PingExample/Collectors/LabelView.swift
@@ -28,7 +28,7 @@ struct LabelView: View {
     @ViewBuilder
     private var labelContent: some View {
         if let richContent = field.richContent {
-            Text(buildAttributedString(from: richContent))
+            Text(RichTextBuilder.build(from: richContent))
         } else {
             Text(field.content)
         }
